This program searches for differences between two image file pictures.
The program scans two image files and compares them by pixels. If it
found differences, it analyzes this fragment of both pictures for
searching the area of mismatch pixels, and places this area in the red
box. First image file is considered as pattern, all red boxes are placed
in second image file.